+++ js/const.js
@@ -1,0 +1,37 @@
+export const PHOTO_COUNT = 25;
+export const AVATAR_COUNT = 6;
+export const MIN_LIKE_COUNT = 15;
+export const MAX_LIKE_COUNT = 200;
+export const COMMENT_COUNT = 17;
+export const NAMES = [
+  'Иван',
+  'Юлия',
+  'Игорь',
+  'Сергей',
+  'София',
+  'Федор',
+  'Геннадий',
+  'Мария',
+];
+
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const DESCRIPTIONS = [
+  'Остаюсь верен традициям – воскресное селфи',
+  'Мыслями на пляже.',
+  'Жизнь слишком коротка для плохого настроения.',
+  'Открываю для себя мир. Скоро вернусь',
+  'Говорю “да” новым приключениям',
+  'Лучшая тренировка – бег на короткие дистанции, от холодильника к телевизору и обратно',
+  'Если бы у меня было чувство юмора, я бы придумал подпись посмешнее',
+  'На календаре четверг воспоминаний. Если оставите комментарий, я на вас подпишусь',
+  'У всех есть такой друг…',
+];
